modified:   include/preyfinder.h 	modified:   src/predator.cpp 	modified:   src/preyfinder.cpp

+++ include/preyfinder.h
@@ -12,7 +12,7 @@
 struct Prey_finder {
     Prey_finder();
     ~Prey_finder();
-    double get_prey(int &);
+    double get_prey(double &);
     double angle;
 private:
     Pixy2  _pixy;
